+++ waspc/cli/src/Wasp/Cli/Command/Call.hs
@@ -17,6 +17,7 @@ data Call
   | Deps
   | Dockerfile
   | Info
+  | Show Arguments
   | News
   | Studio
   | PrintBashCompletionInstruction