+++ waspc/e2e-tests/Main.hs
@@ -29,6 +29,7 @@ import Tests.WaspInfoTest (waspInfoTest)
 import Tests.WaspInstallTest (waspInstallTest)
 import Tests.WaspNewTest (waspNewTest)
 import Tests.WaspProjectLockTest (waspProjectLockTest)
+import Tests.WaspShowTest (waspShowTest)
 import Tests.WaspSpecAvailableTest (waspSpecAvailableTest)
 import Tests.WaspSpecEntityTypesTest (waspSpecEntityTypesTest)
 import Tests.WaspTelemetryTest (waspTelemetryTest)
@@ -109,6 +110,7 @@ e2eTests = do
         waspProjectLockTest,
         waspSpecAvailableTest,
         waspInfoTest,
+        waspShowTest,
         waspInstallTest,
         waspDepsTest,
         waspDockerfileTest,