+++ waspc/src/Wasp/Inspectable.hs
@@ -1,0 +1,22 @@
+module Wasp.Inspectable
+  ( Inspectable (..),
+    InspectionEntry (..),
+    InspectionDatapoint,
+  )
+where
+
+-- | Types that can describe themselves for `wasp show spec`.
+class Inspectable a where
+  inspect :: a -> [InspectionEntry]
+
+data InspectionEntry = InspectionEntry
+  { -- | The category heading for this inspection entry. This is used to group
+    -- related data points together.
+    heading :: String,
+    -- | A list of (label, content) that represent the data points for this
+    -- inspection entry.
+    datapoints :: [InspectionDatapoint]
+  }
+  deriving (Show, Eq)
+
+type InspectionDatapoint = (String, String)